Update README.md for Hacktober_fest2022
Also this year the number to win the swags is reduced to 40000 from 70000 that is being shown in the current README file.  
This is my first time participating in the hacktoberfest, I am not sure if this is how I suggest changes to the README file. Suggestions and feedback will be appreciated.